+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Added `PrepareSandboxTask.pluginName` for easier accessing of the plugin directory name
 - Allow for using non-installer IDEs for plugin verification [#1715](../../issues/1715)
+- Added `bundledModule()` dependency extension helpers
 
 ### Fixed